backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
motor==3.5.1
pymongo==4.8.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.9
selenium==4.24.0

